refactor: consolidate directive utilities
